Using containers instead of „ill“ switch statements and using appropriate data types instead of converting from or to QString

+++ src/modem.cpp
@@ -38,7 +38,7 @@ speed_t Modem::modemspeed() {
 	// convert the string modem speed to a t_speed type
 	// to set the modem.	The constants here should all be ifdef'd because
 	// other systems may not have them
- int i = configdata.getPortSpeed().toInt()/100;
+ int i = configdata.getPortSpeed() /100;
 	switch(i) {
 	case 24:
 	return B2400;
